utils/requirements.txt
livekit-agents
livekit-plugins-turn-detector
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn